[forwarder] limit errors per endpoint per flush
When an endpoint fails 4 times, it is considered down (or at least
unreachable), and all transactions targetting this endpoint are dropped
from the *current* flush. (they are rescheduled for a future flush)

This also changes the way transactions are scheduled in each flush: new
transactions without any error will be tried first.

It also cleans up a little logs.